services/storage-account/func-start-function/requirements.txt
azure-functions
pytest==7.4.2
black==23.11.0
azure.mgmt.subscription==3.1.1
azure.identity==1.14.1
azure.data.tables==12.4.0

python-dotenv==1.0.0
pandas==2.1.1
pytest-cov